requirements.txt
streamlit
PyPDF2
langchain
openai
python-dotenv
langchain-community
langchain-openai
reportlab
pdfkit
markdown2
